small reorder of input files
